+++ MapboxCarPlay/MapboxCarPlay.h
@@ -1,0 +1,9 @@
+#import <UIKit/UIKit.h>
+
+//! Project version number for MapboxCarPlay.
+FOUNDATION_EXPORT double MapboxCarPlayVersionNumber;
+
+//! Project version string for MapboxCarPlay.
+FOUNDATION_EXPORT const unsigned char MapboxCarPlayVersionString[];
+
+