+++ MapboxNavigation/NavigationViewController.swift
@@ -3,9 +3,6 @@ import MapboxCoreNavigation
 import MapboxDirections
 import MapboxSpeech
 import Mapbox
-#if canImport(CarPlay)
-import CarPlay
-#endif
 
 /**
  A ContainerViewController is any UIViewController that conforms to the NavigationComponent messaging protocol.
@@ -110,16 +107,6 @@ open class NavigationViewController: UIViewController {
     @objc public var showsReportFeedback: Bool = true {
         didSet {
             mapViewController?.reportButton.isHidden = !showsReportFeedback
-            showsEndOfRouteFeedback = showsReportFeedback
-        }
-    }
-    
-    /**
-    Shows End of route Feedback UI when the route controller arrives at the final destination. Defaults to `true.`
-    */
-    @objc public var showsEndOfRouteFeedback: Bool = true {
-        didSet {
-            mapViewController?.showsEndOfRoute = showsEndOfRouteFeedback
         }
     }
     
@@ -143,14 +130,6 @@ open class NavigationViewController: UIViewController {
     @objc public var isUsedInConjunctionWithCarPlayWindow = false {
         didSet {
             mapViewController?.isUsedInConjunctionWithCarPlayWindow = isUsedInConjunctionWithCarPlayWindow
-        }
-    }
-    
-    var isConnectedToCarPlay: Bool {
-        if #available(iOS 12.0, *) {
-            return CarPlayManager.isConnected
-        } else {
-            return false
         }
     }
     
@@ -324,49 +303,6 @@ open class NavigationViewController: UIViewController {
         UIApplication.shared.applicationIconBadgeNumber = 1
         UIApplication.shared.applicationIconBadgeNumber = 0
     }
-    
-    #if canImport(CarPlay)
-    /**
-     Presents a `NavigationViewController` on the top most view controller in the window and opens up the `StepsViewController`.
-     If the `NavigationViewController` is already in the stack, it will open the `StepsViewController` unless it is already open.
-     */
-    @available(iOS 12.0, *)
-    public class func carPlayManager(_ carPlayManager: CarPlayManager, didBeginNavigationWith navigationService: NavigationService, window: UIWindow) {
-        
-        if let navigationViewController = window.viewControllerInStack(of: NavigationViewController.self) {
-            // Open StepsViewController on iPhone if NavigationViewController is being presented
-            navigationViewController.isUsedInConjunctionWithCarPlayWindow = true
-        } else {
-            
-            // Start NavigationViewController and open StepsViewController if navigation has not started on iPhone yet.
-            let navigationViewControllerExistsInStack = window.viewControllerInStack(of: NavigationViewController.self) != nil
-            
-            if !navigationViewControllerExistsInStack {
-                
-                let directions = navigationService.directions
-                let route = navigationService.routeProgress.route
-                
-                let service = MapboxNavigationService(route: route, directions: directions, simulating: navigationService.simulationMode)
-                let options = NavigationOptions(navigationService: service)
-                let navigationViewController = NavigationViewController(for: route, options: options)
-                
-                window.rootViewController?.topMostViewController()?.present(navigationViewController, animated: true, completion: {
-                    navigationViewController.isUsedInConjunctionWithCarPlayWindow = true
-                })
-            }
-        }
-    }
-    
-    /**
-     Dismisses a `NavigationViewController` if there is any in the navigation stack.
-     */
-    @available(iOS 12.0, *)
-    public class func carPlayManagerDidEndNavigation(_ carPlayManager: CarPlayManager, window: UIWindow) {
-        if let navigationViewController = window.viewControllerInStack(of: NavigationViewController.self) {
-            navigationViewController.dismiss(animated: true, completion: nil)
-        }
-    }
-    #endif
 }
 
 //MARK: - RouteMapViewControllerDelegate
@@ -521,12 +457,16 @@ extension NavigationViewController: NavigationServiceDelegate {
     @objc public func navigationService(_ service: NavigationService, didArriveAt waypoint: Waypoint) -> Bool {
         let advancesToNextLeg = delegate?.navigationViewController?(self, didArriveAt: waypoint) ?? true
         
-        if !isConnectedToCarPlay, // CarPlayManager shows rating on CarPlay if it's connected
-            service.routeProgress.isFinalLeg && advancesToNextLeg && showsEndOfRouteFeedback {
-            showEndOfRouteFeedback()
+        if service.routeProgress.isFinalLeg {
+            let shouldShowFeedback = delegate?.navigationViewControllerShouldShowEndOfRouteFeedback?(self) ?? true
+            if shouldShowFeedback {
+                defer {
+                    showEndOfRouteFeedback()
+                }
+            }
         }
+        
         return advancesToNextLeg
-
     }
     
     @objc public func showEndOfRouteFeedback(duration: TimeInterval = 1.0, completionHandler: ((Bool) -> Void)? = nil) {